chore: moving db-generator to app directory
